requirements.txt
streamlit>=1.28.1,<2.0.0
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pandas>=2.0.0,<3.0.0
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
plotly>=5.0.0,<6.0.0
